requirements.txt
pillow
transformers
gradio
